+++ lib/src/executable.dart
@@ -182,6 +182,12 @@ transformers:
   } finally {
     await close();
   }
+
+  // TODO(grouma) - figure out why the executable can hang in the travis
+  // environment. https://github.com/dart-lang/test/issues/599
+  if (Platform.environment["FORCE_TEST_EXIT"] == "true") {
+    exit(exitCode);
+  }
 }
 
 /// Print usage information for this command.