+++ lib/src/runner/reporter/expanded.dart
@@ -201,9 +201,7 @@ class ExpandedReporter implements Reporter {
 
     if (error is! LoadException) {
       print(indent(error.toString()));
-      var chain =
-          terseChain(stackTrace, verbose: liveTest.test.metadata.verboseTrace);
-      print(indent(chain.toString()));
+      print(indent('$stackTrace'));
       return;
     }
 
@@ -213,7 +211,7 @@ class ExpandedReporter implements Reporter {
     if (error.innerError is! IsolateSpawnException &&
         error.innerError is! FormatException &&
         error.innerError is! String) {
-      print(indent(terseChain(stackTrace).toString()));
+      print(indent('$stackTrace'));
     }
   }