+++ lib/src/runner/reporter/json.dart
@@ -13,7 +13,6 @@ import '../../backend/state.dart';
 import '../../backend/suite.dart';
 import '../../backend/test_platform.dart';
 import '../../frontend/expect.dart';
-import '../../utils.dart';
 import '../configuration.dart';
 import '../configuration/suite.dart';
 import '../engine.dart';
@@ -250,9 +249,7 @@ class JsonReporter implements Reporter {
     _emit("error", {
       "testID": _liveTestIDs[liveTest],
       "error": error.toString(),
-      "stackTrace":
-          terseChain(stackTrace, verbose: liveTest.test.metadata.verboseTrace)
-              .toString(),
+      "stackTrace": '$stackTrace',
       "isFailure": error is TestFailure
     });
   }